+++ src/TitleDetails/POLDetails/POLDetails.test.js
@@ -1,11 +1,18 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import { render, cleanup } from '@folio/jest-config-stripes/testing-library/react';
 import { IntlProvider } from 'react-intl';
+
+import { render, cleanup } from '@folio/jest-config-stripes/testing-library/react';
+import { useRoutingList } from '@folio/stripes-acq-components';
 
 import '@folio/stripes-acq-components/test/jest/__mock__';
 
 import POLDetails from './POLDetails';
+
+jest.mock('@folio/stripes-acq-components', () => ({
+  ...jest.requireActual('@folio/stripes-acq-components'),
+  useRoutingList: jest.fn(),
+}));
 
 const renderPOLDetails = ({
   accessProvider,
@@ -54,6 +61,9 @@ const polDetails = {
 };
 
 describe('Given POL details', () => {
+  beforeEach(() => {
+    useRoutingList.mockReturnValue({ routingLists: [], isLoading: false });
+  });
   afterEach(cleanup);
 
   it('Than it should display PO Line values', () => {
@@ -67,6 +77,13 @@ describe('Given POL details', () => {
     expect(getByText('ui-receiving.title.orderType.Ongoing')).toBeDefined();
     expect(getByText(polDetails.vendor)).toBeDefined();
     expect(getByText(polDetails.requester)).toBeDefined();
+  });
+
+  it('should display RoutingList component if there is a routingList', () => {
+    useRoutingList.mockReturnValue({ routingLists: ['test'] });
+    const { getByText } = renderPOLDetails(polDetails);
+
+    expect(getByText('ui-receiving.title.hasRouting')).toBeDefined();
   });
 
   describe('When title is connected with poLine', () => {